Final Assignment 02
The task is to classify where a person wearing mask or not.

You have to classify:
	1. Wearing mask
	2. Not wearing mask
For bonus mark you can add more class like:
	# Wearing mask properly
	# Wearing mark not properly

As model you can use:
	# Alexnet or
	# VGG19 or
	# VGG16 or
	# Restnet50 or
	# Any other existing model or
	# Any new model build by you

Finaly you have to use the model in realtime.
You have to use your camera or webcam.
From camera the it will the image file will be taken as input.
From the image file, you have to detect the face using.
For deteting the face, you can use:
	# Medialine or
	# Opencv or
	# Any other library
Taking the detected face as an image, we have to classify using the model you built.